Updated UI, put older files in a folder, removed user data from parser

+++ CPDG_UPDATE.sql
@@ -6,6 +6,6 @@ UPDATE business set numCheckins = checkinTotal.total
 FROM (SELECT businessId, SUM(count) as total FROM checkin GROUP BY businessId) checkinTotal
 WHERE business.businessId = checkinTotal.businessId;
 
-UPDATE business set reviewrating = ratingTotal.starSum/CAST(ratingTotal.count AS decimal(7,1))
+UPDATE business set reviewrating = ratingTotal.starSum/CAST(ratingTotal.count AS FLOAT)
 FROM (SELECT businessId, COUNT(*) as count, SUM(stars) as starSum FROM rating GROUP BY businessId) ratingTotal
 WHERE business.businessId = ratingTotal.businessId;
